refactored some free time functions
