+++ components/EditBeginEndTimeForm.js
@@ -1,31 +1,11 @@
 import { useCallback, useState } from "react";
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
-import FormControl from "react-bootstrap/FormControl";
 import { InputGroup, Container, Row, Col } from "react-bootstrap";
 import useSWR from "swr";
 import { useToasts } from "./Toasts";
 import { Accordion, Card } from "react-bootstrap";
-
-function numToTime(num) {
-  let str = "";
-  if (Math.floor(num / 12) % 12 === 0) {
-    str += "12";
-  } else {
-    str += Math.floor(num / 12) % 12;
-  }
-  str += ":";
-  if ((num % 12) * 5 <= 5) {
-    str += "0";
-  }
-  str += (num % 12) * 5;
-  if (num >= 144) {
-    str += " PM";
-  } else {
-    str += " AM";
-  }
-  return str;
-}
+import numToTime from "../utils/numToTime";
 
 function EditBeginEndTimeForm() {
   const { showToast } = useToasts();
