--- utils/convertTime.js
@@ -1,21 +1,0 @@
-// converts time hh:mm AM/PM to minute of the day
-export default function convertTime(str) {
-  var min = 0;
-  var morn = 0; // 0 if morning, 1 if afternoon
-  if (str.charAt(str.length - 2) == "P") {
-    morn = 1;
-  }
-  str = str.slice(0, -3);
-  var str_split = str.split(":");
-
-  if (str_split[0] == 12) {
-    str_split[0] = 0;
-  }
-
-  min = +str_split[0] * 60 + +str_split[1];
-
-  if (morn == 1) {
-    min = min + 720;
-  }
-  return min;
-}
